big ass changes to project structure
