+++ src/handlers/indexhandler.py
@@ -1,0 +1,4 @@
+import tornado.web
+class IndexHandler(tornado.web.RequestHandler):
+    def get(self):
+        self.render('index.html')
