+++ src/handlers/logouthandler.py
@@ -1,3 +1,4 @@
+import tornado.web
 class LogoutHandler(tornado.web.RequestHandler):
     def get(self):
         self.clear_cookie("user")
