+++ src/server.py
@@ -1,0 +1,21 @@
+import subprocess
+import os.path
+import tornado.httpserver
+import tornado.ioloop
+import tornado.options
+import tornado.web
+from tornado import ioloop, gen
+from tornado.concurrent import Future, chain_future
+from tornado.options import define, options
+from config.routes import application
+
+define("port", default=8000, help="run on the given port", type=int)
+
+def main():
+    tornado.options.parse_command_line()
+    http_server = tornado.httpserver.HTTPServer(application)
+    http_server.listen(options.port)
+    tornado.ioloop.IOLoop.instance().start()
+
+if __name__ == "__main__":
+    main()
