+++ src/templates/index.html
@@ -1,0 +1,5 @@
+<html>
+<body>
+  Hello World!
+</body>
+</html>
